--- services/modules/__init__.py
@@ -1,4 +1,0 @@
-from services.modules.log_module import config_logger
-from services.modules.log_module import logger
-from services.modules.json_validation import validate_request
-from services.modules.preprocessor import preprocess_data
